Merges the writer kernels
